+++ azext_iot/tests/test_iot_ext_unit.py
@@ -42,6 +42,12 @@ device_id = "mydevice"
 child_device_id = "child_device1"
 module_id = "mymod"
 config_id = "myconfig"
+message_etag = "3k28zb44-0d00-4ddd-ade3-6110eb94c476"
+c2d_purge_response = {
+    "deviceId": device_id,
+    "moduleId": None,
+    "totalMessagesPurged": 3
+}
 
 generic_cs_template = "HostName={};SharedAccessKeyName={};SharedAccessKey={}"
 
@@ -1437,7 +1443,7 @@ class TestDeviceModuleMethodInvoke:
 class TestCloudToDeviceMessaging:
     @pytest.fixture(params=["full", "min"])
     def c2d_receive_scenario(self, fixture_ghcs, mocked_response, request):
-        from . generators import create_c2d_receive_response
+        from .generators import create_c2d_receive_response
 
         if request.param == "full":
             payload = create_c2d_receive_response()
@@ -1446,7 +1452,11 @@ class TestCloudToDeviceMessaging:
 
         mocked_response.add(
             method=responses.GET,
-            url=re.compile("https://{}/devices/(.+)/messages/deviceBound".format(mock_target["entity"])),
+            url=re.compile(
+                "https://{}/devices/{}/messages/deviceBound".format(
+                    mock_target["entity"], device_id
+                )
+            ),
             body=payload["body"],
             headers=payload["headers"],
             status=200,
@@ -1454,6 +1464,103 @@ class TestCloudToDeviceMessaging:
         )
 
         yield (mocked_response, payload)
+
+    @pytest.fixture()
+    def c2d_receive_ack_scenario(self, fixture_ghcs, mocked_response):
+        from .generators import create_c2d_receive_response
+        payload = create_c2d_receive_response()
+        mocked_response.add(
+            method=responses.GET,
+            url=(
+                "https://{}/devices/{}/messages/deviceBound".format(
+                    mock_target["entity"], device_id
+                )
+            ),
+            body=payload["body"],
+            headers=payload["headers"],
+            status=200,
+            match_querystring=False,
+        )
+
+        eTag = payload["headers"]["etag"].strip('"')
+        # complete / reject
+        mocked_response.add(
+            method=responses.DELETE,
+            url="https://{}/devices/{}/messages/deviceBound/{}".format(
+                mock_target["entity"], device_id, eTag
+            ),
+            body="",
+            headers=payload["headers"],
+            status=204,
+            match_querystring=False,
+        )
+
+        # abandon
+        mocked_response.add(
+            method=responses.POST,
+            url="https://{}/devices/{}/messages/deviceBound/{}/abandon".format(
+                mock_target["entity"], device_id, eTag
+            ),
+            body="",
+            headers=payload["headers"],
+            status=204,
+            match_querystring=False,
+        )
+
+        yield (mocked_response, payload)
+
+    @pytest.fixture()
+    def c2d_ack_complete_scenario(self, fixture_ghcs, mocked_response):
+        mocked_response.add(
+            method=responses.DELETE,
+            url="https://{}/devices/{}/messages/deviceBound/{}".format(
+                mock_target["entity"], device_id, message_etag
+            ),
+            body="",
+            status=204,
+            match_querystring=False,
+        )
+        yield mocked_response
+
+    @pytest.fixture()
+    def c2d_ack_reject_scenario(self, fixture_ghcs, mocked_response):
+        mocked_response.add(
+            method=responses.DELETE,
+            url="https://{}/devices/{}/messages/deviceBound/{}?reject=".format(
+                mock_target["entity"], device_id, message_etag
+            ),
+            body="",
+            status=204,
+            match_querystring=False,
+        )
+        yield mocked_response
+
+    @pytest.fixture()
+    def c2d_ack_abandon_scenario(self, fixture_ghcs, mocked_response):
+        mocked_response.add(
+            method=responses.POST,
+            url="https://{}/devices/{}/messages/deviceBound/{}/abandon".format(
+                mock_target["entity"], device_id, message_etag
+            ),
+            body="",
+            status=204,
+            match_querystring=False,
+        )
+        yield mocked_response
+
+    @pytest.fixture()
+    def c2d_purge_scenario(self, fixture_ghcs, mocked_response):
+        import json
+        mocked_response.add(
+            method=responses.DELETE,
+            url="https://{}/devices/{}/commands".format(
+                mock_target["entity"], device_id
+            ),
+            body=json.dumps(c2d_purge_response),
+            content_type='application/json',
+            status=200,
+        )
+        yield mocked_response
 
     def test_c2d_receive(self, c2d_receive_scenario):
         service_client = c2d_receive_scenario[0]
@@ -1490,69 +1597,160 @@ class TestCloudToDeviceMessaging:
         if sample_c2d_receive.get("body"):
             assert result["data"] == sample_c2d_receive["body"]
 
-    def test_c2d_complete(self, fixture_service_client_generic):
-        service_client = fixture_service_client_generic
+    def test_c2d_receive_ack(self, c2d_receive_ack_scenario):
+        service_client = c2d_receive_ack_scenario[0]
+        sample_c2d_receive = c2d_receive_ack_scenario[1]
 
-        etag = "3k28zb44-0d00-4ddd-ade3-6110eb94c476"
-        service_client.return_value.status_code = 204
+        timeout = 120
+        for ack in ["complete", "reject", "abandon"]:
+            result = subject.iot_c2d_message_receive(
+                fixture_cmd,
+                device_id,
+                mock_target["entity"],
+                timeout,
+                complete=(ack == 'complete'),
+                reject=(ack == 'reject'),
+                abandon=(ack == 'abandon')
+            )
+            retrieve, action = service_client.calls[0], service_client.calls[1]
+
+            # retrieve call
+            request = retrieve.request
+            url = request.url
+            headers = request.headers
+            assert (
+                "{}/devices/{}/messages/deviceBound?".format(
+                    mock_target["entity"], device_id
+                )
+                in url
+            )
+            assert headers["IotHub-MessageLockTimeout"] == str(timeout)
+
+            assert result["properties"]["system"]["iothub-ack"] == sample_c2d_receive["headers"]["iothub-ack"]
+            assert result["properties"]["system"]["iothub-correlationid"] == sample_c2d_receive["headers"]["iothub-correlationid"]
+            assert result["properties"]["system"]["iothub-deliverycount"] == sample_c2d_receive["headers"]["iothub-deliverycount"]
+            assert result["properties"]["system"]["iothub-expiry"] == sample_c2d_receive["headers"]["iothub-expiry"]
+            assert result["properties"]["system"]["iothub-enqueuedtime"] == sample_c2d_receive["headers"]["iothub-enqueuedtime"]
+            assert result["properties"]["system"]["iothub-messageid"] == sample_c2d_receive["headers"]["iothub-messageid"]
+            assert (
+                result["properties"]["system"]["iothub-sequencenumber"]
+                == sample_c2d_receive["headers"]["iothub-sequencenumber"]
+            )
+            assert result["properties"]["system"]["iothub-userid"] == sample_c2d_receive["headers"]["iothub-userid"]
+            assert result["properties"]["system"]["iothub-to"] == sample_c2d_receive["headers"]["iothub-to"]
+
+            assert result["etag"] == sample_c2d_receive["headers"]["etag"].strip('"')
+
+            if sample_c2d_receive.get("body"):
+                assert result["data"] == sample_c2d_receive["body"]
+
+            # ack call - complete / reject / abandon
+            request = action.request
+            url = request.url
+            headers = request.headers
+            method = request.method
+
+            # all ack calls go to the same base URL
+            assert (
+                "{}/devices/{}/messages/deviceBound/".format(
+                    mock_target["entity"], device_id
+                )
+                in url
+            )
+            # check complete
+            if ack == "complete":
+                assert method == "DELETE"
+            # check reject
+            if ack == "reject":
+                assert method == "DELETE"
+                assert "reject=" in url
+            # check abandon
+            if ack == "abandon":
+                assert method == "POST"
+                assert "/abandon" in url
+            service_client.calls.reset()
+
+    def test_c2d_receive_ack_errors(self):
+        with pytest.raises(CLIError):
+            subject.iot_c2d_message_receive(
+                fixture_cmd,
+                device_id,
+                hub_name=mock_target["entity"],
+                abandon=True,
+                complete=True,
+            )
+            subject.iot_c2d_message_receive(
+                fixture_cmd,
+                device_id,
+                hub_name=mock_target["entity"],
+                abandon=False,
+                complete=True,
+                reject=True,
+            )
+            subject.iot_c2d_message_receive(
+                fixture_cmd,
+                device_id,
+                hub_name=mock_target["entity"],
+                complete=True,
+                reject=True,
+            )
+
+    def test_c2d_complete(self, c2d_ack_complete_scenario):
+        service_client = c2d_ack_complete_scenario
         result = subject.iot_c2d_message_complete(
-            fixture_cmd, device_id, etag, hub_name=mock_target["entity"]
+            fixture_cmd, device_id, message_etag, hub_name=mock_target["entity"]
         )
 
-        args = service_client.call_args
-        url = args[0][0].url
-        method = args[0][0].method
+        request = service_client.calls[0].request
+        url = request.url
+        method = request.method
 
         assert result is None
         assert method == "DELETE"
         assert (
             "{}/devices/{}/messages/deviceBound/{}?".format(
-                mock_target["entity"], device_id, etag
+                mock_target["entity"], device_id, message_etag
             )
             in url
         )
 
-    def test_c2d_reject(self, fixture_service_client_generic):
-        service_client = fixture_service_client_generic
+    def test_c2d_reject(self, c2d_ack_reject_scenario):
+        service_client = c2d_ack_reject_scenario
 
-        etag = "3k28zb44-0d00-4ddd-ade3-6110eb94c476"
-        service_client.return_value.status_code = 204
         result = subject.iot_c2d_message_reject(
-            fixture_cmd, device_id, etag, hub_name=mock_target["entity"]
+            fixture_cmd, device_id, message_etag, hub_name=mock_target["entity"]
         )
 
-        args = service_client.call_args
-        url = args[0][0].url
-        method = args[0][0].method
+        request = service_client.calls[0].request
+        url = request.url
+        method = request.method
 
         assert result is None
         assert method == "DELETE"
         assert (
             "{}/devices/{}/messages/deviceBound/{}?".format(
-                mock_target["entity"], device_id, etag
+                mock_target["entity"], device_id, message_etag
             )
             in url
         )
         assert "reject=" in url
 
-    def test_c2d_abandon(self, fixture_service_client_generic):
-        service_client = fixture_service_client_generic
+    def test_c2d_abandon(self, c2d_ack_abandon_scenario):
+        service_client = c2d_ack_abandon_scenario
 
-        etag = "3k28zb44-0d00-4ddd-ade3-6110eb94c476"
-        service_client.return_value.status_code = 204
         result = subject.iot_c2d_message_abandon(
-            fixture_cmd, device_id, etag, hub_name=mock_target["entity"]
+            fixture_cmd, device_id, message_etag, hub_name=mock_target["entity"]
         )
 
-        args = service_client.call_args
-        url = args[0][0].url
-        method = args[0][0].method
+        request = service_client.calls[0].request
+        url = request.url
+        method = request.method
 
         assert result is None
         assert method == "POST"
         assert (
             "{}/devices/{}/messages/deviceBound/{}/abandon?".format(
-                mock_target["entity"], device_id, etag
+                mock_target["entity"], device_id, message_etag
             )
             in url
         )
@@ -1571,6 +1769,21 @@ class TestCloudToDeviceMessaging:
             subject.iot_c2d_message_reject(
                 fixture_cmd, device_id, hub_name=mock_target["entity"], etag=""
             )
+
+    def test_c2d_message_purge(self, c2d_purge_scenario):
+        result = subject.iot_c2d_message_purge(fixture_cmd, device_id)
+        request = c2d_purge_scenario.calls[0].request
+        url = request.url
+        method = request.method
+
+        assert method == "DELETE"
+        assert "https://{}/devices/{}/commands".format(
+            mock_target["entity"], device_id
+        ) in url
+        assert result
+        assert result.total_messages_purged == 3
+        assert result.device_id == device_id
+        assert not result.module_id
 
 
 class TestSasTokenAuth: